Implement design into more pages
