+++ frontend/design-system/button/text-button/text-button.tsx
@@ -6,17 +6,22 @@ import { ReactNode } from 'react';
 import styles from './text-button.module.scss';
 
 type ButtonProps = {
+    color?: string;
     startIcon?: ReactNode;
     endIcon?: ReactNode;
     disabled?: boolean;
     fullWidth?: boolean;
 } & React.HTMLAttributes<HTMLButtonElement>;
 
-export const TextButton: React.FC<ButtonProps> = ({ children, startIcon, endIcon, disabled, fullWidth, ...rest }) => {
+export const TextButton: React.FC<ButtonProps> = ({ children, color, startIcon, endIcon, disabled, fullWidth, ...rest }) => {
     const { theme, toggleTheme } = useTheme();
 
+    const colors = {
+        color: color ? 'var(--' + color + ')' : undefined
+    };
+
     return (
-        <button className={classNames(styles.baseButton, fullWidth ? styles.fullWidth : '')} disabled={disabled} aria-disabled={disabled} {...rest}>
+        <button style={colors} className={classNames(styles.baseButton, fullWidth ? styles.fullWidth : '')} disabled={disabled} aria-disabled={disabled} {...rest}>
             {startIcon && <Box className={styles.iconWrapper}>{startIcon}</Box>}
             <Label variant="button">{children}</Label>
             {endIcon && <Box className={styles.iconWrapper}>{endIcon}</Box>}
