--- frontend/pages/game.module.scss
@@ -1,110 +1,0 @@
-.question {
-    &__title {
-        font-family: $font-4;
-        font-size: 3rem;
-        padding: rem(20);
-        width: 100%;
-    }
-}
-
-.answers {
-    display: flex;
-    flex-direction: column;
-    gap: rem(20);
-    width: 100%;
-    align-items: center;
-    padding: rem(20);
-
-    &__answer {
-        background-color: $onyx;
-        border: none;
-        color: $white;
-        border-radius: 5px;
-        padding: rem(10);
-        width: 100%;
-        cursor: pointer;
-        font-family: $font-1;
-    }
-}
-
-// Result background wrapper
-.result {
-    width: 100%;
-    height: 100%;
-}
-
-.host {
-    display: flex;
-    flex-direction: column;
-
-    &__question {
-        font-family: $font-4;
-        font-size: 3rem;
-        font-weight: 700;
-    }
-
-    &__players {
-        display: flex;
-        flex-direction: column;
-        gap: rem(20);
-        padding: rem(20) 0;
-
-        &__player {
-            width: 100%;
-            background-color: $gray-200;
-            color: $white;
-            border-radius: 5px;
-
-            padding: rem(10) rem(20);
-
-            display: flex;
-            flex-direction: row;
-            align-items: center;
-
-            &__score {
-                margin-left: auto;
-                background-color: $gray-100;
-                padding: rem(5) rem(10);
-                border-radius: 5px;
-            }
-        }
-    }
-}
-
-.header {
-    position: relative;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    nav {
-        width: 100%;
-        padding: 1rem;
-
-        ul {
-            display: flex;
-            justify-content: flex-start;
-            align-items: center;
-            gap: rem(40);
-            margin: 0;
-            padding: 0;
-
-            li {
-                margin: auto 0;
-                list-style: none;
-                font-size: 1.125rem;
-                position: relative;
-                padding: 1.5rem 0;
-
-                // Current player score
-                &.score {
-                    margin-left: auto;
-                }
-
-                &.name {
-                    font-weight: 700;
-                }
-            }
-        }
-    }
-}
