--- frontend/pages/waiting.module.scss
@@ -1,61 +1,0 @@
-.waiting {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    &__info {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        text-align: center;
-        width: 50%;
-
-        &__title {
-            font-family: $font-4;
-        }
-
-        &__id {
-            color: $secondary-100;
-            font-size: 4rem;
-        }
-    }
-
-    &__players {
-        width: 100%;
-        display: flex;
-        flex-direction: row;
-        margin-top: rem(40);
-        padding: rem(10);
-
-        flex-wrap: wrap;
-
-        gap: 1rem;
-    }
-
-    // Backround display
-    &__player {
-        display: flex;
-        flex-direction: row;
-        justify-content: flex-start;
-        padding: rem(20);
-        background-color: $gray-200;
-        border-radius: 10px;
-        color: $white;
-
-        flex: 40%;
-        max-width: 50%;
-
-        align-items: center;
-
-        // Admin edit options
-        &__kick {
-            display: flex;
-            flex-direction: row;
-            margin-left: auto;
-            align-items: center;
-            color: $white;
-            cursor: pointer;
-        }
-    }
-}
