+++ frontend/pages/waiting/waiting.module.scss
@@ -1,0 +1,12 @@
+.kickButton {
+    cursor: pointer;
+}
+
+.players {
+    width: 100%;
+}
+
+.playerItem {
+    flex: 40%;
+    max-width: 50%;
+}
